Sista fixet samt lite ändringar på att registrerar sig och spara information om den user som registrerar sig.
