+++ app/src/main/AndroidManifest.xml
@@ -14,12 +14,27 @@
         tools:targetApi="31">
         <activity
 
+
             android:name=".AddNChangeFoodActivity"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
+        <activity
+            android:name=".RegisterActivity"
+            android:exported="false">
+            <meta-data
+            android:name="android.app.lib_name"
+            android:value=""
+            /></activity>
+        <activity
+            android:name=".LoginAndRegisterActivity"
+            android:exported="false">
+            <meta-data
+                android:name="android.app.lib_name"
+                android:value=""
+                /></activity>
         <activity
