+++ app/src/main/res/layout/activity_login_and_register.xml
@@ -12,23 +12,25 @@
         android:id="@+id/feedMeLogoImageView"
         android:layout_width="140dp"
         android:layout_height="139dp"
-        app:layout_constraintBottom_toBottomOf="parent"
+        android:layout_marginBottom="46dp"
+        app:layout_constraintBottom_toTopOf="@+id/logoImageView"
         app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.498"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.047"
+        app:layout_constraintVertical_bias="0.0"
         app:srcCompat="@drawable/logo" />
 
     <EditText
         android:id="@+id/emailTextView"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:background="@color/white"
+        android:background="@drawable/border"
         android:ems="10"
-        android:hint="email"
+        android:hint="  email"
         android:inputType="textPersonName"
         android:minHeight="48dp"
-        android:textColor="#FFFFFF"
+        android:textColor="#000000"
         android:textColorHint="#959595"
         android:visibility="visible"
         app:layout_constraintBottom_toBottomOf="parent"
@@ -42,9 +44,9 @@
         android:id="@+id/passwordTextView"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:background="@color/white"
+        android:background="@drawable/border"
         android:ems="10"
-        android:hint="password"
+        android:hint="  password"
         android:inputType="textPersonName"
         android:minHeight="48dp"
         android:textColorHint="#A5A5A5"
