+++ app/src/main/res/layout/activity_register.xml
@@ -23,11 +23,11 @@
         android:layout_height="46dp"
         android:layout_marginTop="16dp"
         android:layout_marginBottom="14dp"
-        android:background="@color/white"
+        android:background="@drawable/border"
         android:ems="10"
-        android:hint="Förnamn"
+        android:hint="  Förnamn"
         android:inputType="textPersonName"
-        android:textColor="#FFFFFF"
+        android:textColor="#000000"
         android:textColorHint="#8F8F8F"
         app:layout_constraintBottom_toTopOf="@+id/efterNamnTextView"
         app:layout_constraintEnd_toEndOf="parent"
@@ -40,11 +40,11 @@
         android:layout_height="46dp"
         android:layout_marginTop="76dp"
         android:layout_marginBottom="10dp"
-        android:background="@color/white"
+        android:background="@drawable/border"
         android:ems="10"
-        android:hint="Efternamn"
+        android:hint="  Efternamn"
         android:inputType="textPersonName"
-        android:textColor="#FFFFFF"
+        android:textColor="#000000"
         android:textColorHint="#8F8F8F"
         app:layout_constraintBottom_toTopOf="@+id/emailTextView1"
         app:layout_constraintEnd_toEndOf="parent"
@@ -57,66 +57,51 @@
         android:layout_width="211dp"
         android:layout_height="46dp"
         android:layout_marginTop="132dp"
-        android:layout_marginBottom="10dp"
-        android:background="@color/white"
+        android:background="@drawable/border"
         android:ems="10"
-        android:hint="Email"
+        android:hint="  Email"
         android:inputType="textPersonName"
-        android:textColor="#FFFFFF"
+        android:textColor="#000000"
         android:textColorHint="#8F8F8F"
-        app:layout_constraintBottom_toTopOf="@+id/editTextTextPersonName4"
+        app:layout_constraintBottom_toTopOf="@+id/passwordTextView1"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.505"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/iV_foodDisplay_RV" />
 
     <EditText
-        android:id="@+id/editTextTextPersonName4"
-        android:layout_width="211dp"
-        android:layout_height="46dp"
-        android:layout_marginTop="10dp"
-        android:background="@color/white"
-        android:ems="10"
-        android:hint="Email"
-        android:inputType="textPersonName"
-        android:textColor="#FFFFFF"
-        android:textColorHint="#8F8F8F"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.505"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/emailTextView1" />
-
-    <EditText
         android:id="@+id/passwordTextView1"
         android:layout_width="211dp"
         android:layout_height="46dp"
         android:layout_marginTop="10dp"
-        android:background="@color/white"
+        android:layout_marginBottom="15dp"
+        android:background="@drawable/border"
         android:ems="10"
-        android:hint="Password"
+        android:hint="  Password"
         android:inputType="textPersonName"
-        android:textColor="#FFFFFF"
+        android:textColor="#000000"
         android:textColorHint="#8F8F8F"
+        app:layout_constraintBottom_toTopOf="@+id/whatUserTextView"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.51"
+        app:layout_constraintHorizontal_bias="0.505"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/emailTextView1" />
+        app:layout_constraintTop_toBottomOf="@+id/emailTextView1"
+        app:layout_constraintVertical_bias="0.0" />
 
     <EditText
         android:id="@+id/userTextView"
         android:layout_width="211dp"
         android:layout_height="46dp"
         android:layout_marginTop="104dp"
-        android:layout_marginBottom="14dp"
-        android:background="@color/white"
+        android:background="@drawable/border"
         android:ems="10"
-        android:hint="Admin, Customer, Courier?"
-        android:textSize="16sp"
+        android:hint="  Admin, Customer, Courier?"
         android:inputType="textPersonName"
-        android:textColor="#FFFFFF"
+        android:textColor="#000000"
         android:textColorHint="#8F8F8F"
-        app:layout_constraintBottom_toTopOf="@+id/registerButton1"
+        android:textSize="16sp"
         app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.505"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/emailTextView1" />
 
@@ -161,50 +146,34 @@
         app:layout_constraintVertical_bias="0.546" />
 
     <Button
-        android:id="@+id/registerButton1"
-        android:layout_width="197dp"
-        android:layout_height="51dp"
-        android:layout_marginStart="102dp"
-        android:layout_marginTop="14dp"
-        android:layout_marginEnd="112dp"
-        android:layout_marginBottom="208dp"
-        android:text="Register"
-        android:textColor="#FFFFFF"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/userTextView" />
-
-    <Button
         android:id="@+id/saveButton"
         android:layout_width="197dp"
         android:layout_height="51dp"
-        android:layout_marginStart="102dp"
-        android:layout_marginEnd="112dp"
+        android:layout_marginStart="100dp"
+        android:layout_marginTop="20dp"
+        android:layout_marginEnd="100dp"
         android:layout_marginBottom="156dp"
         android:text="Save User"
         android:textColor="#FFFFFF"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="1.0"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/registerButton1"
-        app:layout_constraintVertical_bias="1.0" />
+        app:layout_constraintTop_toBottomOf="@+id/userTextView"
+        app:layout_constraintVertical_bias="0.0" />
 
     <TextView
         android:id="@+id/whatUserTextView"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginStart="30dp"
-        android:layout_marginTop="16dp"
         android:layout_marginEnd="160dp"
-        android:layout_marginBottom="13dp"
+        android:layout_marginBottom="8dp"
         android:text="Type of User?"
         android:textColor="#FFFFFF"
         app:layout_constraintBottom_toTopOf="@+id/userTextView"
         app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.533"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/editTextTextPersonName4" />
+        app:layout_constraintTop_toBottomOf="@+id/passwordTextView1" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
